Renamed googlehome.ts to chromecast.ts
